# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RequestUploadUrlBody {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: RequestUploadUrlBody;
}

export interface ErrorEnvelope {
  error: string;
}

export interface OnboardingRequest {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  email: string;
  phone?: string;
  brokerage?: string;
  /** @minLength 1 */
  mlsAgentId: string;
  personalWebsiteUrl?: string;
  headshotUrl?: string;
  logoUrl?: string;
}

export interface OnboardingResponse {
  agentId: string;
  /** Token-bearing URL for fetching/updating the agent's profile (e.g. `/api/agents/profile?token=…`). */
  profileUrl: string;
  /** Stripe Checkout setup-mode URL. Redirect the browser here when present. */
  checkoutUrl?: string | null;
  /** How many existing preview listings were linked to this new agent by `listAgentMlsId`. */
  backfilledCount?: number | null;
}

export interface AgentProfile {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  phone?: string | null;
  brokerage?: string | null;
  mlsAgentId: string;
  personalWebsiteUrl?: string | null;
  headshotUrl?: string | null;
  logoUrl?: string | null;
}

export interface AgentProfileResponse {
  agent: AgentProfile;
}

export type AnalyticsEventInputEventType =
  (typeof AnalyticsEventInputEventType)[keyof typeof AnalyticsEventInputEventType];

export const AnalyticsEventInputEventType = {
  pageview: "pageview",
  session_start: "session_start",
  session_end: "session_end",
  gallery_photo_view: "gallery_photo_view",
  lead_submitted: "lead_submitted",
} as const;

export interface AnalyticsEventInput {
  listingId: string;
  /**
   * @minLength 8
   * @maxLength 64
   */
  sessionId: string;
  eventType: AnalyticsEventInputEventType;
  occurredAt?: string;
  /** @maxLength 2000 */
  referrer?: string | null;
  /** @maxLength 100 */
  utmSource?: string | null;
  /**
   * @minimum 0
   * @maximum 500
   */
  photoIndex?: number | null;
  /** @maxLength 500 */
  path?: string | null;
}

export interface IngestAnalyticsEventsRequest {
  /**
   * @minItems 1
   * @maxItems 50
   */
  events: AnalyticsEventInput[];
}

export interface CreateLeadRequest {
  listingId: string;
  /**
   * @minLength 1
   * @maxLength 120
   */
  name: string;
  /** @maxLength 200 */
  email: string;
  /** @maxLength 40 */
  phone?: string;
  /** @maxLength 2000 */
  message?: string;
  /** @maxLength 40 */
  source?: string;
}

/**
 * All fields optional. Send `null` for `headshotUrl`/`logoUrl`/`personalWebsiteUrl` to clear that field.
 */
export interface AgentProfileUpdate {
  /** @minLength 1 */
  firstName?: string;
  /** @minLength 1 */
  lastName?: string;
  phone?: string;
  brokerage?: string;
  personalWebsiteUrl?: string | null;
  headshotUrl?: string | null;
  logoUrl?: string | null;
}

/**
 * Missing or invalid agent token.
 */
export type UnauthorizedResponse = ErrorEnvelope;

/**
 * Resource not found.
 */
export type NotFoundResponse = ErrorEnvelope;

/**
 * Invalid request.
 */
export type BadRequestResponse = ErrorEnvelope;

export type OnboardAgent200 = {
  /** True when the agent's MLS prefix is outside the served market. */
  outOfMarket: boolean;
};

export type GetAgentProfileParams = {
  token: string;
};

export type UpdateAgentProfileParams = {
  token: string;
};

export type GetAgentBillingPortalParams = {
  token: string;
};

export type GetAgentBillingPortal200 = {
  url: string;
};

export type ActivateListingParams = {
  token: string;
};

export type ActivateListing200 = {
  success?: boolean;
  already?: boolean;
  domainName?: string | null;
  subscriptionId?: string | null;
  replitHandoffNote?: string | null;
};

export type CreateLead201 = {
  leadId: string;
};

export type EmailUnsubscribeOneClickParams = {
  email: string;
  /**
   * HMAC-signed token bound to this email address.
   */
  token: string;
};

export type EmailUnsubscribePostBodyOne = {
  email: string;
  token: string;
};

export type EmailUnsubscribePostBodyTwo = {
  email: string;
  token: string;
};

export type EmailUnsubscribePost200 = {
  ok?: boolean;
};

export type BackfillWeeklyReportBody = {
  /** UTC instant corresponding to local Monday 00:00 of the target week. */
  weekStart?: string;
};

export type BackfillWeeklyReport200 = {
  ok: boolean;
  sent?: boolean;
  weekStart: string;
};

export type BackfillWeeklyReport409 = {
  ok?: boolean;
  weekStart?: string;
  reason?: string;
};