# lib/api-zod/src/generated/types/analyticsEventInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsEventInputEventType } from "./analyticsEventInputEventType";

export interface AnalyticsEventInput {
  listingId: string;
  /**
   * @minLength 8
   * @maxLength 64
   */
  sessionId: string;
  eventType: AnalyticsEventInputEventType;
  occurredAt?: Date;
  /** @maxLength 2000 */
  referrer?: string | null;
  /** @maxLength 100 */
  utmSource?: string | null;
  /**
   * @minimum 0
   * @maximum 500
   */
  photoIndex?: number | null;
  /** @maxLength 500 */
  path?: string | null;
}