# lib/api-zod/src/generated/types/ingestAnalyticsEventsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsEventInput } from "./analyticsEventInput";

export interface IngestAnalyticsEventsRequest {
  /**
   * @minItems 1
   * @maxItems 50
   */
  events: AnalyticsEventInput[];
}